+++ src/IssueViz/Editor/SelectedIssueTagging/Adornment/IssueLocationAdornmentTagger.cs
@@ -18,18 +18,14 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
-using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Text.Editor;
 using Microsoft.VisualStudio.Text.Tagging;
-using SonarLint.VisualStudio.IssueVisualization.Models;
 
 namespace SonarLint.VisualStudio.IssueVisualization.Editor.SelectedIssueTagging.Adornment
 {
-    internal sealed class IssueLocationAdornmentTagger : FilteringTaggerBase<ISelectedIssueLocationTag, IntraTextAdornmentTag>
+    internal sealed partial class IssueLocationAdornmentTagger : FilteringTaggerBase<ISelectedIssueLocationTag, IntraTextAdornmentTag>
     {
         private readonly IWpfTextView wpfView;
         private readonly ICachingAdornmentFactory adornmentCache;
@@ -57,7 +53,7 @@ namespace SonarLint.VisualStudio.IssueVisualization.Editor.SelectedIssueTagging.
             // for all of its tags.
             var wholeSpan = new SnapshotSpan(wpfView.TextSnapshot, 0, wpfView.TextSnapshot.Length);
             var allTags = tagAggregator.GetTags(wholeSpan);
-            adornmentCache.RemoveUnused(allTags.Select(x => x.Tag.Location));
+            adornmentCache.Refresh(allTags.Select(x => x.Tag.Location));
         }
 
         protected override TagSpan<IntraTextAdornmentTag> CreateTagSpan(ISelectedIssueLocationTag trackedTag, NormalizedSnapshotSpanCollection spans)
@@ -79,72 +75,5 @@ namespace SonarLint.VisualStudio.IssueVisualization.Editor.SelectedIssueTagging.
             }
             base.Dispose(disposing);
         }
-
-        #region Adornment cache
-
-        internal interface ICachingAdornmentFactory
-        {
-            /// <summary>
-            /// Clears unnecessary entries from the cache
-            /// </summary>
-            /// <param name="currentLocations">The list of locations that are still in use</param>
-            void RemoveUnused(IEnumerable<IAnalysisIssueLocationVisualization> currentLocations);
-
-            IssueLocationAdornment CreateOrUpdate(IAnalysisIssueLocationVisualization locationViz);
-        }
-
-        internal class CachingAdornmentFactory : ICachingAdornmentFactory
-        {
-            private readonly IWpfTextView wpfTextView;
-            private readonly IDictionary<IAnalysisIssueLocationVisualization, IssueLocationAdornment> locVizToAdornmentMap;
-
-            internal /* for testing */ IReadOnlyCollection<IssueLocationAdornment> CachedAdornments => locVizToAdornmentMap.Values.ToList();
-
-            public CachingAdornmentFactory(IWpfTextView view)
-            {
-                wpfTextView = view;
-                locVizToAdornmentMap = new Dictionary<IAnalysisIssueLocationVisualization, IssueLocationAdornment>();
-            }
-
-            public IssueLocationAdornment CreateOrUpdate(IAnalysisIssueLocationVisualization locationViz)
-            {
-                // As long as the cache is always accessed on the same thread we don't need to worry
-                // about synchronising access to it. Currently VS is always running this code on the
-                // main thread. The assertion is to detect if this behaviour changes.
-                Debug.Assert(ThreadHelper.CheckAccess(), "Expected cache to be accessed on the UI thread");
-
-                if (locVizToAdornmentMap.TryGetValue(locationViz, out var existingAdornment))
-                {
-                    existingAdornment.Update(wpfTextView.FormattedLineSource);
-                    return existingAdornment;
-                }
-
-                var newAdornment = CreateAdornment(locationViz);
-                locVizToAdornmentMap[locationViz] = newAdornment;
-                return newAdornment;
-            }
-
-            public void RemoveUnused(IEnumerable<IAnalysisIssueLocationVisualization> currentLocations)
-            {
-                Debug.Assert(ThreadHelper.CheckAccess(), "Expected cache to be accessed on the UI thread");
-
-                var unused = locVizToAdornmentMap.Keys.Except(currentLocations).ToArray();
-                foreach(var item in unused)
-                {
-                    locVizToAdornmentMap.Remove(item);
-                }
-            }
-
-            private IssueLocationAdornment CreateAdornment(IAnalysisIssueLocationVisualization locationViz)
-            {
-                var adornment = new IssueLocationAdornment(locationViz, wpfTextView.FormattedLineSource);
-
-                // If we don't call Measure here the tag is positioned incorrectly
-                adornment.Measure(new System.Windows.Size(double.PositiveInfinity, double.PositiveInfinity));
-                return adornment;
-            }
-        }
-
-        #endregion Adornment cache
     }
 }
